+++ index.js
@@ -1,54 +1,46 @@
-// @ts-check
-
-import Bluebird from "bluebird"
+import "moment-timezone"
 import "artsy-newrelic"
+import Bluebird from "bluebird"
 import xapp from "artsy-xapp"
-import cors from "cors"
-import depthLimit from "graphql-depth-limit"
-import morgan from "artsy-morgan"
 import express from "express"
 import forceSSL from "express-force-ssl"
-import graphqlHTTP from "express-graphql"
 import bodyParser from "body-parser"
-import { mergeSchemas } from "./lib/mergeSchemas"
-import localSchema from "./schema"
-import legacyLoaders from "./lib/loaders/legacy"
-import createLoaders from "./lib/loaders"
-import config from "./config"
-import { info, error } from "./lib/loggers"
-import graphqlErrorHandler from "./lib/graphql-error-handler"
-import moment from "moment"
-import "moment-timezone"
-import { fetchLoggerSetup, fetchLoggerRequestDone } from "lib/loaders/api/logger"
-import { middleware as requestTracer, makeSchemaTraceable } from "./lib/tracer"
-import { middleware as requestIDsAdder } from "./lib/requestIDs"
+import config from "./src/config"
+import { createReloadable } from "@artsy/express-reloadable"
+import { info, error } from "./src/lib/loggers"
+
+const { GRAVITY_API_URL, GRAVITY_ID, GRAVITY_SECRET, NODE_ENV, PORT } = process.env
 
 global.Promise = Bluebird
 
-const {
-  PORT,
-  NODE_ENV,
-  GRAVITY_API_URL,
-  GRAVITY_ID,
-  GRAVITY_SECRET,
-  QUERY_DEPTH_LIMIT,
-  ENABLE_QUERY_TRACING,
-  ENABLE_SCHEMA_STITCHING,
-} = process.env
-
 const app = express()
 const port = PORT || 3000
-const queryLimit = (QUERY_DEPTH_LIMIT && parseInt(QUERY_DEPTH_LIMIT, 10)) || 10 // Default to ten.
-
+const isDevelopment = NODE_ENV === "development"
 const isProduction = NODE_ENV === "production"
-const enableQueryTracing = ENABLE_QUERY_TRACING === "true"
-const enableSchemaStitching = ENABLE_SCHEMA_STITCHING === "true"
 
-function startApp(schema) {
+xapp.on("error", err => {
+  error(err)
+  process.exit()
+})
+
+const xappConfig = {
+  url: GRAVITY_API_URL,
+  id: GRAVITY_ID,
+  secret: GRAVITY_SECRET,
+}
+
+xapp.init(xappConfig, () => {
+  config.GRAVITY_XAPP_TOKEN = xapp.token
+  bootApp() // eslint-disable-line
+})
+
+function bootApp() {
   if (isProduction) {
     app.set("forceSSLOptions", { trustXFPHeader: true }).use(forceSSL)
     app.set("trust proxy", 1)
   }
+
+  app.use(bodyParser.json())
 
   app.get("/favicon.ico", (_req, res) => {
     res
@@ -59,85 +51,154 @@ function startApp(schema) {
 
   app.all("/graphql", (_req, res) => res.redirect("/"))
 
-  app.use(bodyParser.json())
-
-  if (enableQueryTracing) {
-    console.warn("[FEATURE] Enabling query tracing") // eslint-disable-line
-    makeSchemaTraceable(schema)
-    app.use(requestTracer)
+  if (isDevelopment) {
+    const mountAndReload = createReloadable(app, require)
+    mountAndReload("./src")
+  } else {
+    app.use(require("./src").default)
   }
-
-  app.use(requestIDsAdder)
-
-  app.use(
-    "/",
-    cors(),
-    morgan,
-    graphqlHTTP((req, res) => {
-      info("----------")
-
-      legacyLoaders.clearAll()
-
-      const accessToken = req.headers["x-access-token"]
-      const userID = req.headers["x-user-id"]
-      const timezone = req.headers["x-timezone"]
-
-      const { requestIDs, span } = res.locals
-      const requestID = requestIDs.requestID
-
-      if (!isProduction) {
-        fetchLoggerSetup(requestID)
-      }
-
-      // Accepts a tz database timezone string. See http://www.iana.org/time-zones,
-      // https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
-      let defaultTimezone
-      if (moment.tz.zone(timezone)) {
-        defaultTimezone = timezone
-      }
-
-      const loaders = createLoaders(accessToken, userID, requestID)
-      // Share with e.g. the Convection ApolloLink in mergedSchema.
-      res.locals.dataLoaders = loaders // eslint-disable-line no-param-reassign
-
-      return {
-        schema,
-        graphiql: true,
-        rootValue: {
-          accessToken,
-          userID,
-          defaultTimezone,
-          span,
-          ...createLoaders(accessToken, userID, requestIDs),
-        },
-        formatError: graphqlErrorHandler(req.body),
-        validationRules: [depthLimit(queryLimit)],
-        extensions: isProduction ? undefined : fetchLoggerRequestDone(requestID),
-      }
-    })
-  )
 
   app.listen(port, () => info(`Listening on ${port}`))
 }
 
-xapp.on("error", err => {
-  error(err)
-  process.exit()
-})
+// import Bluebird from "bluebird"
+// import "artsy-newrelic"
+// import xapp from "artsy-xapp"
+// import cors from "cors"
+// import depthLimit from "graphql-depth-limit"
+// import morgan from "artsy-morgan"
+// import express from "express"
+// import forceSSL from "express-force-ssl"
+// import graphqlHTTP from "express-graphql"
+// import bodyParser from "body-parser"
+// import { mergeSchemas } from "./lib/mergeSchemas"
+// import localSchema from "./schema"
+// import legacyLoaders from "./lib/loaders/legacy"
+// import createLoaders from "./lib/loaders"
+// import config from "./config"
+// import { info, error } from "./lib/loggers"
+// import graphqlErrorHandler from "./lib/graphql-error-handler"
+// import moment from "moment"
+// import "moment-timezone"
+// import { fetchLoggerSetup, fetchLoggerRequestDone } from "lib/loaders/api/logger"
+// import { middleware as requestTracer, makeSchemaTraceable } from "./lib/tracer"
+// import { middleware as requestIDsAdder } from "./lib/requestIDs"
 
-xapp.init(
-  {
-    url: GRAVITY_API_URL,
-    id: GRAVITY_ID,
-    secret: GRAVITY_SECRET,
-  },
-  () => {
-    config.GRAVITY_XAPP_TOKEN = xapp.token
-    if (enableSchemaStitching) {
-      console.warn("[FEATURE] Enabling schema stitching") // eslint-disable-line
-      mergeSchemas().then(startApp)
-    } else {
-      startApp(localSchema)
-    }
-  }
-)
+// global.Promise = Bluebird
+
+// const {
+//   PORT,
+//   NODE_ENV,
+//   GRAVITY_API_URL,
+//   GRAVITY_ID,
+//   GRAVITY_SECRET,
+//   QUERY_DEPTH_LIMIT,
+//   ENABLE_QUERY_TRACING,
+//   ENABLE_SCHEMA_STITCHING,
+// } = process.env
+
+// const app = express()
+// const port = PORT || 3000
+// const queryLimit = (QUERY_DEPTH_LIMIT && parseInt(QUERY_DEPTH_LIMIT, 10)) || 10 // Default to ten.
+
+// const isProduction = NODE_ENV === "production"
+// const enableQueryTracing = ENABLE_QUERY_TRACING === "true"
+// const enableSchemaStitching = ENABLE_SCHEMA_STITCHING === "true"
+
+// function startApp(schema) {
+//   if (isProduction) {
+//     app.set("forceSSLOptions", { trustXFPHeader: true }).use(forceSSL)
+//     app.set("trust proxy", 1)
+//   }
+
+//   app.get("/favicon.ico", (_req, res) => {
+//     res
+//       .status(200)
+//       .set({ "Content-Type": "image/x-icon" })
+//       .end()
+//   })
+
+//   app.all("/graphql", (_req, res) => res.redirect("/"))
+
+//   app.use(bodyParser.json())
+
+//   if (enableQueryTracing) {
+//     console.warn("[FEATURE] Enabling query tracing") // eslint-disable-line
+//     makeSchemaTraceable(schema)
+//     app.use(requestTracer)
+//   }
+
+//   app.use(requestIDsAdder)
+
+//   app.use(
+//     "/",
+//     cors(),
+//     morgan,
+//     graphqlHTTP((req, res) => {
+//       info("----------")
+
+//       legacyLoaders.clearAll()
+
+//       const accessToken = req.headers["x-access-token"]
+//       const userID = req.headers["x-user-id"]
+//       const timezone = req.headers["x-timezone"]
+
+//       const { requestIDs, span } = res.locals
+//       const requestID = requestIDs.requestID
+
+//       if (!isProduction) {
+//         fetchLoggerSetup(requestID)
+//       }
+
+//       // Accepts a tz database timezone string. See http://www.iana.org/time-zones,
+//       // https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+//       let defaultTimezone
+//       if (moment.tz.zone(timezone)) {
+//         defaultTimezone = timezone
+//       }
+
+//       const loaders = createLoaders(accessToken, userID, requestID)
+//       // Share with e.g. the Convection ApolloLink in mergedSchema.
+//       res.locals.dataLoaders = loaders // eslint-disable-line no-param-reassign
+
+//       return {
+//         schema,
+//         graphiql: true,
+//         rootValue: {
+//           accessToken,
+//           userID,
+//           defaultTimezone,
+//           span,
+//           ...createLoaders(accessToken, userID, requestIDs),
+//         },
+//         formatError: graphqlErrorHandler(req.body),
+//         validationRules: [depthLimit(queryLimit)],
+//         extensions: isProduction ? undefined : fetchLoggerRequestDone(requestID),
+//       }
+//     })
+//   )
+
+//   app.listen(port, () => info(`Listening on ${port}`))
+// }
+
+// xapp.on("error", err => {
+//   error(err)
+//   process.exit()
+// })
+
+// xapp.init(
+//   {
+//     url: GRAVITY_API_URL,
+//     id: GRAVITY_ID,
+//     secret: GRAVITY_SECRET,
+//   },
+//   () => {
+//     config.GRAVITY_XAPP_TOKEN = xapp.token
+//     if (enableSchemaStitching) {
+//       console.warn("[FEATURE] Enabling schema stitching") // eslint-disable-line
+//       mergeSchemas().then(startApp)
+//     } else {
+//       startApp(localSchema)
+//     }
+//   }
+// )
