+++ package.json
@@ -106,11 +106,11 @@
     "typescript": "^2.5.2"
   },
   "jest": {
-    "setupTestFrameworkScriptFile": "<rootDir>/test/helper.js",
+    "setupTestFrameworkScriptFile": "<rootDir>/src/test/helper.js",
     "testPathIgnorePatterns": [
       "/node_modules/",
       "/build/",
-      "/test/helper.js",
+      "/src/test/helper.js",
       "/src/test/utils.js",
       "/src/test/gql.js",
       "/src/test/__mocks__"
